Rake task for adding Monroe Park codes, and make external downloading conditional on files not existing in /tmp

+++ lib/tasks/property_data.rake
@@ -3,13 +3,18 @@ require 'csv'
 namespace :property_data do
   desc "Import data from the property data file and create an addresses JSON file for typeahead search"
   task :import => :environment do
-    p "Downloading property data..."
-    system("cd /tmp && curl -O https://s3-us-west-1.amazonaws.com/south-bend-secrets/Vacant_and_Abandoned_Property_Data.csv")
-    p "Downloading centroids..."
-    system("cd /tmp && curl -O https://s3-us-west-1.amazonaws.com/south-bend-secrets/cityparcelscentroids_abandoned_latlon_CLEAN.csv")
     property_data_path = "/tmp/Vacant_and_Abandoned_Property_Data.csv"
+    unless File.exist?(property_data_path)
+      p "Downloading property data..."
+      system("cd /tmp && curl -O https://s3-us-west-1.amazonaws.com/south-bend-secrets/Vacant_and_Abandoned_Property_Data.csv")
+    end
+    centroid_path = "/tmp/cityparcelscentroids_abandoned_latlon_CLEAN.csv"
+    unless File.exist?(centroid_path)
+      p "Downloading centroids..."
+      system("cd /tmp && curl -O https://s3-us-west-1.amazonaws.com/south-bend-secrets/cityparcelscentroids_abandoned_latlon_CLEAN.csv")
+    end
     table = CSV.read(property_data_path, :headers => true)
-    lat_long_table = CSV.read("/tmp/cityparcelscentroids_abandoned_latlon_CLEAN.csv", :headers => true)
+    lat_long_table = CSV.read(centroid_path, :headers => true)
     all_address_array = []
     lats_and_longs_array = []
     if ENV["MONROE_PILOT"]
@@ -25,6 +30,9 @@ namespace :property_data do
       address = row["Location 1"][0, row["Location 1"].index("\n")]
       # Remove periods from addresses
       clean_address = address.gsub(".", "")
+      if clean_address == "520 Columbia" || clean_address == "620 Columbia"
+        clean_address.gsub!("Columbia", "Carroll")
+      end
       all_address_array << clean_address
       if target_property == nil
         target_property = Property.create(:parcel_id => parcel_id, :name => clean_address)
@@ -68,9 +76,26 @@ namespace :property_data do
     File.delete(lats_and_longs_array_path) if File.exist?(lats_and_longs_array_path)
     File.open(lats_and_longs_array_path, 'w') { |file| file.write(lats_and_longs_array.to_json) }
   end
+
   desc "PENDING - Pull down CSV data from Socrata and store in /tmp"
   task :download_from_socrata do
     # Pending
+  end
+
+  desc "Adds for Monroe Park"
+  task :add_monroe_phone_codes => :environment do
+    monroe_park_codes = { "519 S St Joseph" => "2345", "523 S St Joseph" => "3456", "213 E South" => "4567", "614 S St Joseph" => "5678", "520 Carroll" => "6789", "620 Carroll" => "7891", "615 Fellows" => "8912", "624 Fellows" => "9123", "616 Clinton" => "1234" }
+    # Test to check they're all good
+    #monroe_park_codes.each_pair { |key,value| p "uh oh: #{key} #{value}" if Property.find_all_by_name(key).count == 0 }
+    monroe_park_codes.each_pair do |prop_address, prop_code|
+     target = Property.find_by_name(prop_address)
+     if target.property_code
+       p "#{target.name} already has a property code"
+     else
+       target.update_attribute(:property_code, prop_code)
+       p "Added property code #{prop_code} to #{target.name}"
+     end
+    end
   end
 end
